% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od-funs.R
\name{od2odf}
\alias{od2odf}
\title{Extract coordinates from OD data}
\usage{
od2odf(flow, zones)
}
\description{
Extract coordinates from OD data
}
\details{
Origin-destination (OD) data is often provided
in the form of 1 line per OD pair, with zone codes of the trip origin in the first
column and the zone codes of the destination in the second column
(see the [`vignette("stplanr-od")`](https://docs.ropensci.org/stplanr/articles/stplanr-od.html)) for details.
`od2odf()` creates an 'origin-destination data frame', based on a data frame containing
origin and destination cones (`flow`) that match the first column in a
a spatial (polygon or point) object (`zones`).

The function returns a data frame with coordinates for the origin and destination.
}
\examples{
data(flow)
data(zones)
od2odf(flow[1:2, ], zones)
}
