% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od-funs.R
\name{od_coordinates_from_points}
\alias{od_coordinates_from_points}
\title{# @examples
# x = od_data_leeds
# p = sf::st_centroid(od::od_data_zones)}
\usage{
od_coordinates_from_points(x, p, p_destinations = NULL)
}
\description{
# @examples
# x = od_data_leeds
# p = sf::st_centroid(od::od_data_zones)
}
