% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od-funs.R
\name{od_coordinates_ids}
\alias{od_coordinates_ids}
\title{Interleave origin and destination coordinates}
\usage{
od_coordinates_ids(odc)
}
\arguments{
\item{odc}{A matrix containing coordinates representing line start and end points}
}
\description{
This function takes a matrix with 4 columns representing origin and destination coordinates
and returns a data frame with 3 columns with the ID of each linestring, plus
the coordinates representing origin and destination coordinates.
Essentially the function is a pivot, converting from wide to long format, to feed
into other functions for creating geographic desire lines.
}
\examples{
od_coordinates_ids(od_coordinates(od_data_df, p = od_data_zones, sfnames = TRUE))
}
