% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{od_data_buildings}
\alias{od_data_buildings}
\title{Simple buildings dataset}
\description{
Building data from OSM for testing od_disaggregate.
}
\examples{
nrow(od_data_buildings)
head(od_data_buildings)
plot(od_data_buildings$geometry)
plot(od_data_zones_min$geometry, lwd = 3, col = NULL, add = TRUE)
}
\keyword{datasets}
