% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{od_data_centroids}
\alias{od_data_centroids}
\alias{od_data_coordinates}
\title{Datasets reporesenting zone centroids}
\description{
These are provided as a geographic (sf) object and a simple
data frame with longitude (\code{X}) and latitude (\code{Y}) columns.
}
\note{
The schema data can be (re-)generated using code in the
\code{data-raw} directory.
}
\examples{
od_data_centroids
head(od_data_coordinates)
}
\keyword{datasets}
