% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{od_data_centroids2}
\alias{od_data_centroids2}
\title{Output area centroids}
\description{
This dataset represents geographic centroids of Output Areas in Leeds, UK.

This dataset represents geographic centroids of Output Areas in Leeds, UK.
}
\note{
The schema data can be (re-)generated using code in the
\code{data-raw} directory.

The schema data can be (re-)generated using code in the
\code{data-raw} directory.
}
\examples{
head(od_data_centroids2)
head(od_data_centroids2)
}
\keyword{datasets}
