% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{od_data_destinations}
\alias{od_data_destinations}
\title{Workplace zone (destination) centroids}
\description{
This dataset represents geographic centroids of Output Areas in Leeds, UK.
}
\note{
The schema data can be (re-)generated using code in the
\code{data-raw} directory.
}
\examples{
nrow(od_data_destinations)
head(od_data_destinations)
}
\keyword{datasets}
