% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{od_data_df}
\alias{od_data_df}
\title{Minimal origin-destination dataset}
\description{
This is a minimal dataset representing top commuter desire lines
in Leeds based on the 2011 Census.
The first two variables of the data frame are the zone code of origin and destination, respectively.
The other columns record the number of people who travel by different modes, including
`all`, `train`, `bus`, `bicycle` and by `foot`.
}
\note{
The schema data can be (re-)generated using code in the
`data-raw` directory.
}
\keyword{datasets}
