% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{od_data_df2}
\alias{od_data_df2}
\title{Origin-destination data with destinations in a different layer than origins}
\description{
This dataset represents commuter flows between Output Areas
and Workplace Zones, the most detailed open OD data in the UK.
See https://wicid.ukdataservice.ac.uk/ and the script
\code{data-raw/od_wpz.R} in the \code{od} package's GitHub repo.
}
\details{
The dataset reports (in the 3rd column) the number of people travelling
between origins and destinations.
}
\note{
The schema data can be (re-)generated using code in the
\code{data-raw} directory.
}
\examples{
head(od_data_df2)
}
\keyword{datasets}
