% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{od_data_leeds}
\alias{od_data_leeds}
\alias{od_data_csa_zones}
\alias{od_data_zones}
\title{Example od data}
\description{
Small OD datasets for packages examples
}
\note{
The schema data can be (re-)generated using code in the
`data-raw` directory.
}
\keyword{datasets}
