% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{od_data_network}
\alias{od_data_network}
\title{Route network data for Leeds}
\description{
Route network data for Leeds
}
\note{
The schema data can be (re-)generated using code in the
`data-raw` directory.
}
\examples{
od_data_network
head(od_data_network)
}
\keyword{datasets}
