% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{od_data_zones}
\alias{od_data_zones}
\alias{od_data_csa_zones}
\alias{od_data_zones_min}
\title{Example OD data}
\description{
Zone datasets for packages examples
}
\note{
The schema data can be (re-)generated using code in the
\code{data-raw} directory.
}
\keyword{datasets}
