% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{od_data_zones_small}
\alias{od_data_zones_small}
\title{Small zones dataset}
\description{
This dataset represents geographic zones of Lower Super Output Areas in Leeds, UK.
They fit completely within the \code{od_data_zones_min} dataset.
}
\note{
The schema data can be (re-)generated using code in the
\code{data-raw} directory.
}
\examples{
nrow(od_data_zones_small)
head(od_data_zones_small)
plot(od_data_zones_small$geometry)
plot(od_data_zones_min$geometry, lwd = 3, col = NULL, add = TRUE)
}
\keyword{datasets}
