% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points_to_od.R
\name{od_interzone}
\alias{od_interzone}
\alias{od_intrazone}
\title{Return only interzonal (io intrazonal) OD pairs}
\usage{
od_interzone(x)

od_intrazone(x)
}
\arguments{
\item{x}{A data frame in which the first two columns are codes
representing points/zones of origin and destination}
}
\description{
This function takes an OD dataset and returns only the rows
corresponding to movements in which the origin is different than
the destination.
}
\examples{

od_data = points_to_od(od_data_centroids)
nrow(od_data)
nrow(od_interzone(od_data))
nrow(od_intrazone(od_data))
}
