% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jitter.R
\name{od_jitter}
\alias{od_jitter}
\title{Move desire line end points within zone to avoid all trips going to a single centroid}
\usage{
od_jitter(od, z, zd = NULL, subpoints = NULL, subpoints_d = NULL)
}
\arguments{
\item{od}{An origin-destination data frame}

\item{z}{Zones representing origins and destinations}

\item{zd}{Zones with ids matching the destination codes in input OD data}

\item{subpoints}{Points within the zones defining the OD data start/end points}

\item{subpoints_d}{Points withing destination zones representing possible destinations}
}
\value{
An \code{sf} data frame
}
\description{
These functions tackle the problem associated with OD data representing movement to and from large zones.
Typically the associated desire lines start and end in one point per zone.
This function produces desire lines that can start and end anywhere (or at predefined points) within each zone.
See \href{https://github.com/ITSLeeds/od/issues/11}{issue #11} for details.
}
\examples{

# Basic example
od = od_data_df
z = od_data_zones_min
desire_lines_rand = od_jitter(od, z)
desire_lines = od_to_sf(od, z)
plot(z$geometry)
plot(desire_lines_rand$geometry, add = TRUE)
plot(desire_lines, add = TRUE)

# Example showing use of subpoints
subpoints = sf::st_sample(z, 200)
subpoints_d = sf::st_sample(z, 100)
desire_lines_rand_d = od_jitter(od, z, subpoints = subpoints, subpoints_d = subpoints_d)
plot(z$geometry)
plot(desire_lines_rand_d$geometry, add = TRUE)
plot(subpoints, add = TRUE)
plot(subpoints_d, col = "red", add = TRUE)
plot(desire_lines, add = TRUE)
# mapview::mapview(desire_lines) + desire_lines_rand + z # interactive map
subpoints = sf::st_sample(z, 100)
desire_lines_rand2 = od_jitter(desire_lines, z, subpoints = subpoints)
plot(z, reset = FALSE)
plot(subpoints, add = TRUE)
plot(desire_lines_rand2$geometry, add = TRUE)

# Example showing jittering of origins and destinations
od = od_data_df2
z = sf::st_buffer(od_data_centroids2, dist = 1000)
zd = sf::st_buffer(od_data_destinations, dist = 300)
zd = zd[zd[[1]] \%in\% od[[2]], ]
desire_lines = od_to_sf(od, od_data_centroids2, od_data_destinations)
desire_lines_rand = od_jitter(od, z, zd)
plot(z$geometry)
plot(od_data_centroids2$geometry, add = TRUE)
plot(od_data_destinations$geometry, add = TRUE)
plot(zd$geometry, add = TRUE)
plot(desire_lines$geometry, add = TRUE)
plot(desire_lines_rand$geometry, add = TRUE)

# Larger example with only subset of matching zones
# od = od_data_df_medium
# od_sf = od_to_sf(od, od_data_zones)
# desire_lines_rand3 = od_jitter(od_sf, z)
# plot(od_sf$geometry[od$all > 200])
# plot(desire_lines_rand3$geometry[od$all > 200])
# mapview::mapview(od_sf$geometry[od$all > 200])
}
