% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{od_road_network}
\alias{od_road_network}
\title{Simple road network dataset}
\description{
Road network data from OSM for testing od_disaggregate.
}
\examples{
library(sf)
nrow(od_road_network)
head(od_road_network)
plot(od_road_network$geometry)
plot(od_data_zones_min$geometry, lwd = 3, col = NULL, add = TRUE)
}
\keyword{datasets}
