% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{od_sample_vertices}
\alias{od_sample_vertices}
\title{Create a subsample of points from a route network for jittering}
\usage{
od_sample_vertices(x, fraction = 1)
}
\arguments{
\item{x}{An sf object representing a road network}

\item{fraction}{What percent of the network to sample?}
}
\description{
Todo: export this at some point
}
\examples{
\dontrun{
u = "https://github.com/ITSLeeds/od/releases/download/v0.3.1/road_network_min.Rds"
f = basename(u)
if (!file.exists(f)) download.file(u, f)
road_network_min = readRDS(f)
od_sample_vertices(road_network_min)
}
}
