% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od-funs.R
\name{od_to_odmatrix}
\alias{od_to_odmatrix}
\title{Convert origin-destination data from long to wide format}
\usage{
od_to_odmatrix(flow, attrib = 3, name_orig = 1, name_dest = 2)
}
\arguments{
\item{flow}{A data frame representing flows between origin and destinations}

\item{attrib}{A number or character string representing the column containing the attribute data
of interest from the `flow` data frame}

\item{name_orig}{A number or character string representing the zone of origin}

\item{name_dest}{A number or character string representing the zone of destination}
}
\description{
This function takes a data frame representing travel between origins
(with origin codes in `name_orig`, typically the 1st column)
and destinations
(with destination codes in `name_dest`, typically the second column) and returns a matrix
with cell values (from `attrib`, the third column by default) representing travel between
origins and destinations.
}
\examples{
# od_to_odmatrix(flow)
# od_to_odmatrix(flow[1:9, ])
# od_to_odmatrix(flow[1:9, ], attrib = "Bicycle")
}
\seealso{
Other od: 
\code{\link{odmatrix_to_od}()}
}
\concept{od}
