% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od-funs.R
\name{od_to_sf}
\alias{od_to_sf}
\title{Convert origin-destination coordinates into desire lines}
\usage{
od_to_sf(odf, zones, destinations = NULL)
}
\arguments{
\item{odc}{A data frame or matrix representing the coordinates
of origin-destination data. The first two columns represent the
coordinates of the origin (typically longitude and latitude) points;
the third and fourth columns represent the coordinates of the destination
(in the same CRS). Each row represents travel from origin to destination.}

\item{crs}{A number representing the coordinate reference system
of the result, 4326 by default.}
}
\description{
Convert origin-destination coordinates into desire lines
}
\examples{
desire_lines = od_to_sf(od_data_leeds, od_data_zones)
plot(desire_lines)

# desire_lines2 = stplanr::od2line(od_data_leeds, od_data_zones)
# plot(desire_lines2[-c(1:2)])
# bench::mark(check = FALSE,
#   desire_lines = od_to_sf(od_data_leeds, od_data_zones)
#   desire_lines2 = stplanr::od2line(od_data_leeds, od_data_zones)
# )
}
