% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od-funs.R
\name{od_to_sfc}
\alias{od_to_sfc}
\title{Convert origin-destination coordinates into geographic 'desire line' objects}
\usage{
od_to_sfc(x, z, zd = NULL, verbose = FALSE)
}
\arguments{
\item{x}{A data frame in which the first two columns are codes
representing points/zones of origin and destination}

\item{z}{Zones representing origins and destinations}

\item{zd}{Zones representing destinations}

\item{verbose}{Print messages providing progress updates? FALSE by default}
}
\description{
Convert origin-destination coordinates into geographic 'desire line' objects
}
\examples{
desire_lines = od_to_sfc(od_data_leeds, od_data_zones)
desire_lines[1:3]
if(requireNamespace("stplanr") && requireNamespace("bench")) {
desire_lines2 = stplanr::od2line(od_data_leeds, od_data_zones)
plot(desire_lines)
plot(desire_lines2)
bench::mark(check = FALSE, max_iterations = 10,
  od = od_to_sfc(od_data_leeds, od_data_zones),
  stplanr = stplanr::od2line(od_data_leeds, od_data_zones)
)
}
}
