% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od-funs.R
\name{odc_to_sf}
\alias{odc_to_sf}
\title{Convert od coordinates to sf object}
\usage{
odc_to_sf(odc, df, package = "sfheaders", crs = 4326)
}
\description{
Convert od coordinates to sf object
}
\examples{
odc = od_coordinates(od_data_leeds, z = od_data_zones)
odc
# odl = od_coordinates_to_sfc(odc) # currently fails...
# class(odl)
# plot(odl)

# odc_to_sf(odc, df = od_data_leeds)
# odsf = odc_to_sf()

# testing the result
# odlsf = od_coordinates_to_sfc(odc, "sf")
# odsf_stplanr = stplanr::od2line(od_data_leeds, od_data_zones)
# odl_stplanr = sf::st_geometry(odsf_stplanr)
# odl_stplanr[1]
# odl[1]
# odlsf[1]
# bench::mark(check = FALSE,
#   od_coordinates_to_sfc(odc),
#   od_coordinates_to_sfc(odc, package = "sf"),
#   stplanr::od2line(od_data_leeds, od_data_zones)
# )
# system.time({odl_stplanr = stplanr::od2line(od_data_leeds, od_data_zones)})
# plot(sf::st_geometry(odl_stplanr))
}
