% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od-funs.R
\name{odc_to_sfc}
\alias{odc_to_sfc}
\title{Convert origin-destination coordinates into geographic desire lines}
\usage{
odc_to_sfc(odc)
}
\arguments{
\item{odc}{A matrix containing coordinates representing line start and end points}
}
\description{
Convert origin-destination coordinates into geographic desire lines
}
\examples{
(odc = od_coordinates(od_data_df, p = od::od_data_zones, sfnames = TRUE))
(l = odc_to_sfc(odc))
plot(l)
}
