% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points_to_od.R
\name{points_to_od}
\alias{points_to_od}
\title{Convert a series of points into a dataframe of origins and destinations}
\usage{
points_to_od(p)
}
\arguments{
\item{p}{A spatial points object}
}
\description{
Takes a series of geographical points and converts them into a data.frame
representing the potential flows, or 'spatial interaction', between every combination
of points.
}
\examples{
p = od_data_centroids[1:3, ]
od1 <- points_to_od(p)
od1
od1$v = 1
unique(od1)
od_oneway(od1)
}
