% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od-funs.R
\name{od_coordinates}
\alias{od_coordinates}
\title{Create matrices representing origin-destination coordinates}
\usage{
od_coordinates(x, p = NULL, silent = TRUE, sfnames = FALSE)
}
\arguments{
\item{x}{A data frame in which the first two columns are codes
representing points/zones of origin and destination}

\item{p}{Points representing origins and destinations}

\item{silent}{Hide messages? \code{FALSE} by default.}

\item{sfnames}{Should output column names be compatible with the sf package?}
}
\description{
This function takes a wide range of input data types (spatial lines, points or text strings)
and returns a matrix of coordinates representing origin (ox, oy) and destination (dx, dy) points.
}
\examples{
x = od_data_df
p = od_data_centroids
od_coordinates(x, p)[1:2, ]
od_coordinates(x, p, sfnames = TRUE)[1:2, ]
od_coordinates(x, p, silent = FALSE)[1:2, ]
# x[[1]][1] =  "404"
# Next line will error:
# od_coordinates(x, p, silent = FALSE)[1:2, ]
# From original stplanr function:
# od_coords(from = c(0, 52), to = c(1, 53)) # lon/lat coordinates
# od_coords(from = cents[1, ], to = cents[2, ]) # Spatial points
# od_coords(cents_sf[1:3, ], cents_sf[2:4, ]) # sf points
# # od_coords("Hereford", "Leeds") # geocode locations
# od_coords(flowlines[1:3, ])
# od_coords(flowlines_sf[1:3, ])
}
