% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{od_disaggregate}
\alias{od_disaggregate}
\title{Split-up each OD pair into multiple OD pairs based on subpoints/subzones}
\usage{
od_disaggregate(od, z, subpoints = NULL, subzones = NULL, code_append = "_ag")
}
\arguments{
\item{od}{An origin-destination data frame}

\item{z}{Zones representing origins and destinations}

\item{subpoints}{Points within the zones defining the OD data}

\item{subzones}{Sub-zones within the zones defining the OD data}

\item{code_append}{The name of the column containing aggregate zone names}
}
\description{
This function is for splitting-up OD pairs, it's roughly and OD data equivalent
of the \code{tidyr::pivot_longer()} function.
}
\examples{
od = od::od_data_df
zones = od::od_data_zones_min
subzones = od_data_zones_small
od_disag = od_disaggregate(od, zones, subzones)
ncol(od_disag) -1 == ncol(od) # same number of columns (except disag data gained geometry)
sum(od_disag[[3]]) == sum(od[[3]])
sum(od_disag[[4]]) == sum(od[[4]])
od_sf = od_to_sf(od, zones)
plot(od_data_zones_small$geometry)
plot(od_data_zones_min$geometry, lwd = 3, col = NULL, add = TRUE)
plot(od_sf["all"], add = TRUE)
plot(od_disag["all"], add = TRUE)
}
