% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od-funs.R
\name{od_to_sf}
\alias{od_to_sf}
\title{Convert origin-destination coordinates into desire lines}
\usage{
od_to_sf(x, z, zd = NULL)
}
\arguments{
\item{x}{A data frame in which the first two columns are codes
representing points/zones of origin and destination}

\item{z}{Zones representing origins and destinations}

\item{zd}{Zones representing destinations}
}
\description{
Convert origin-destination coordinates into desire lines
}
\examples{
desire_lines = od_to_sf(od_data_leeds, od_data_zones)
plot(desire_lines)

# desire_lines2 = stplanr::od2line(od_data_leeds, od_data_zones)
# plot(desire_lines2[-c(1:2)])
# bench::mark(check = FALSE,
#   desire_lines = od_to_sf(od_data_leeds, od_data_zones)
#   desire_lines2 = stplanr::od2line(od_data_leeds, od_data_zones)
# )
}
