% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od-funs.R
\name{odc_to_sf}
\alias{odc_to_sf}
\title{Convert origin-destination coordinates into geographic desire lines}
\usage{
odc_to_sf(odc, d = NULL, crs = 4326)
}
\arguments{
\item{odc}{A matrix containing coordinates representing line start and end points}

\item{d}{An optional data frame to add to the geometry column}

\item{crs}{The coordinate reference system of the output, if not known in \code{z}.
4326 by default.}
}
\description{
Convert origin-destination coordinates into geographic desire lines
}
\examples{
(odc = od_coordinates(od_data_df, p = od_data_zones, sfnames = TRUE))
(l = odc_to_sf(odc))
plot(l)
lsfc = odc_to_sfc(odc)
}
