% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od-funs.R
\name{odc_to_sf}
\alias{odc_to_sf}
\title{Convert origin-data coordinates into sf object}
\usage{
odc_to_sf(odc, df, package = "sfheaders", crs = 4326)
}
\arguments{
\item{odc}{A data frame or matrix representing the coordinates
of origin-destination data. The first two columns represent the
coordinates of the origin (typically longitude and latitude) points;
the third and fourth columns represent the coordinates of the destination
(in the same CRS). Each row represents travel from origin to destination.
Convert od coordinates to sf object}

\item{df}{Data frame with OD data attributes}

\item{package}{Which package should create the geographic object? `sfheaders` by default.}

\item{crs}{The coordinate reference system of the output, 4326 by default}
}
\description{
Convert origin-data coordinates into sf object
}
\examples{
# testing the result
# odlsf = od_coordinates_to_sfc(odc, "sf")
# odsf_stplanr = stplanr::od2line(od_data_leeds, od_data_zones)
# odl_stplanr = sf::st_geometry(odsf_stplanr)
# odl_stplanr[1]
# odl[1]
# odlsf[1]
# bench::mark(check = FALSE,
#   od_coordinates_to_sfc(odc),
#   od_coordinates_to_sfc(odc, package = "sf"),
#   stplanr::od2line(od_data_leeds, od_data_zones)
# )
# system.time({odl_stplanr = stplanr::od2line(od_data_leeds, od_data_zones)})
# plot(sf::st_geometry(odl_stplanr))
}
